+++ c-ext/compressoriterator.c
@@ -57,14 +57,8 @@ feedcompressor:
 	/* If we have data left in the input, consume it. */
 	if (self->input.pos < self->input.size) {
 		Py_BEGIN_ALLOW_THREADS
-		if (self->compressor->mtcctx) {
-			zresult = ZSTDMT_compressStream(self->compressor->mtcctx,
-				&self->output, &self->input);
-		}
-		else {
-			zresult = ZSTD_compressStream(self->compressor->cstream, &self->output,
-				&self->input);
-		}
+		zresult = ZSTD_compress_generic(self->compressor->cctx, &self->output,
+			&self->input, ZSTD_e_continue);
 		Py_END_ALLOW_THREADS
 
 		/* Release the Python object holding the input buffer. */
@@ -129,12 +123,12 @@ feedcompressor:
 
 	/* EOF */
 	if (0 == readSize) {
-		if (self->compressor->mtcctx) {
-			zresult = ZSTDMT_endStream(self->compressor->mtcctx, &self->output);
-		}
-		else {
-			zresult = ZSTD_endStream(self->compressor->cstream, &self->output);
-		}
+		self->input.src = NULL;
+		self->input.size = 0;
+		self->input.pos = 0;
+
+		zresult = ZSTD_compress_generic(self->compressor->cctx, &self->output,
+			&self->input, ZSTD_e_end);
 		if (ZSTD_isError(zresult)) {
 			PyErr_Format(ZstdError, "error ending compression stream: %s",
 				ZSTD_getErrorName(zresult));
@@ -158,13 +152,8 @@ feedcompressor:
 	self->input.pos = 0;
 
 	Py_BEGIN_ALLOW_THREADS
-	if (self->compressor->mtcctx) {
-		zresult = ZSTDMT_compressStream(self->compressor->mtcctx, &self->output,
-			&self->input);
-	}
-	else {
-		zresult = ZSTD_compressStream(self->compressor->cstream, &self->output, &self->input);
-	}
+	zresult = ZSTD_compress_generic(self->compressor->cctx, &self->output,
+		&self->input, ZSTD_e_continue);
 	Py_END_ALLOW_THREADS
 
 	/* The input buffer currently points to memory managed by Python